FINAL VERSION 1.0.0
Final version where i added the ability to search postcode and search for closest charging station, added ability to add products to cart and checkout, then saved to order history, added ability to cancel or edit the booking details. as well added some validation on input fields and for some pages to load only if user is logged in.